+++ hata/ext/commands_v2/context.py
@@ -344,10 +344,8 @@ class CommandContext(object):
             if (failed_check is not None):
                 raise CommandCheckError(failed_check)
             
-            wrappers = self.command._wrappers
-            if (wrappers is not None):
-                for wrapper in wrappers:
-                    await wrapper(self)
+            for wrapper in command_function._iter_wrappers():
+                await wrapper(self)
             
             parameter_parsing_states = await command_function._content_parser.parse_content(self, index)
             command_positional_parameters = self.command_positional_parameters